working on eliminating some warnings, now that I set the warning level to -Wall

+++ src/DnnLayerConvolution.cpp
@@ -90,37 +90,37 @@ namespace tfs {
         if( m_in_a == 0 || m_out_a == 0 ) {
             return log_error( "Not configured" );
         }
-        const long in_x   = m_in_a->width();        // var V_sx = V.sx |0;
-        const long in_y   = m_in_a->height();       // var V_sy = V.sy |0;
-        const long in_z   = m_in_a->depth();
-        const long side   = m_side;
-        const long stride = m_stride;              // var xy_stride = this.stride |0;
+        const unsigned long in_x   = m_in_a->width();        // var V_sx = V.sx |0;
+        const unsigned long in_y   = m_in_a->height();       // var V_sy = V.sy |0;
+        const unsigned long in_z   = m_in_a->depth();
+        const unsigned long side   = m_side;
+        const unsigned long stride = m_stride;              // var xy_stride = this.stride |0;
         
-        const long out_x = m_out_a->width();
-        const long out_y = m_out_a->height();
-        const long out_z = m_out_a->depth();       // Filter count
+        const unsigned long out_x = m_out_a->width();
+        const unsigned long out_y = m_out_a->height();
+        const unsigned long out_z = m_out_a->depth();       // Filter count
 
-        const long last = side-1;
-        for( long d = 0; d < out_z; d++ ) {        // Filter count
-            long x = -m_pad;
-            for( long ax = 0; ax < out_x; x += stride, ax++ ) {
-                long y = -m_pad;
-                for( long ay = 0; ay < out_y; y += stride, ay++ ) {
+        const unsigned long last = side-1;
+        for( unsigned long d = 0; d < out_z; d++ ) {        // Filter count
+            long x = - (long) m_pad;
+            for( unsigned long ax = 0; ax < out_x; x += stride, ax++ ) {
+                long y = - (long) m_pad;
+                for( unsigned long ay = 0; ay < out_y; y += stride, ay++ ) {
                     // convolve centered at this point
                     DNN_NUMERIC a = 0.0;
-                    for( long fx = 0; fx < side; fx++ ) {
-                        for( long  fy = 0; fy < side; fy++ ) {
-                            for( long fd = 0;fd < in_z; fd++ ) {
-                                long oy = y+fy; // coordinates in the original input array coordinates
-                                long ox = x+fx;
-                                if( oy >= 0 && oy < in_y && ox >= 0 && ox < in_x ) {
-                                    a += m_w->get( d, fx, fy, fd ) * m_in_a->get( ox, oy, fd );
+                    for( unsigned long fx = 0; fx < side; fx++ ) {
+                        for( unsigned long  fy = 0; fy < side; fy++ ) {
+                            for( unsigned long fd = 0;fd < in_z; fd++ ) {
+                                long oy = y + (long) fy; // coordinates in the original input array coordinates
+                                long ox = x + (long) fx;
+                                if( oy >= 0 && oy < (long) in_y && ox >= 0 && ox < (long) in_x ) {
+                                    a += m_w->get( d, fx, fy, fd ) * m_in_a->get((unsigned long) ox, (unsigned long)oy, fd );
                                 }
                             }
                         }
                     }
                     a += m_w->get( d, last, last, in_z );   // bias
-                    m_out_a->set( ax, ay, d, a );       // set output
+                    m_out_a->set( ax, ay, d, a );           // set output
                 }
             }
         }
@@ -135,41 +135,41 @@ namespace tfs {
         if( m_in_a == 0 || m_w == 0 || m_dw == 0 ) {
             return log_error( "Not configured" );
         }
-        const long in_x   = m_in_a->width();        // var V_sx = V.sx |0;
-        const long in_y   = m_in_a->height();       // var V_sy = V.sy |0;
-        const long in_z   = m_in_a->depth();
-        const long side   = m_side;
-        const long stride = m_stride;              // var xy_stride = this.stride |0;
+        const unsigned long in_x   = m_in_a->width();        // var V_sx = V.sx |0;
+        const unsigned long in_y   = m_in_a->height();       // var V_sy = V.sy |0;
+        const unsigned long in_z   = m_in_a->depth();
+        const unsigned long side   = m_side;
+        const unsigned long stride = m_stride;              // var xy_stride = this.stride |0;
         
-        const long out_x = m_out_dw->width();
-        const long out_y = m_out_dw->height();
-        const long out_z = m_out_dw->depth();
+        const unsigned long out_x = m_out_dw->width();
+        const unsigned long out_y = m_out_dw->height();
+        const unsigned long out_z = m_out_dw->depth();
 
         if( m_in_dw != 0 ) {        // Input layer often does not have dw.
             m_in_dw->zero();        // Zero input gradiant, we add to it below.
         }
-        const long last = side-1;
-        for( long d = 0; d < out_z; d++ ) {
-            long x = -m_pad;
-            for( long ax = 0; ax < out_x; x += stride, ax++ ) {
-                long y = -m_pad;
-                for( long ay = 0; ay < out_y; y += stride, ay++ ) {
+        const unsigned long last = side-1;
+        for( unsigned long d = 0; d < out_z; d++ ) {
+            long x = -(long) m_pad;
+            for( unsigned long ax = 0; ax < out_x; x += stride, ax++ ) {
+                long y = -(long) m_pad;
+                for( unsigned long ay = 0; ay < out_y; y += stride, ay++ ) {
                     // convolve and add up the gradients.
                     const DNN_NUMERIC chain_grad = m_out_dw->get( ax, ay, d ); // gradient from chain rule
-                    for( long fx = 0; fx < side; fx++ ) {
-                        for( long  fy = 0; fy < side; fy++ ) {
-                            for( long fd = 0;fd < in_z; fd++ ) {
-                                long oy = y+fy; // coordinates in the original input array coordinates
-                                long ox = x+fx;
-                                if( oy >= 0 && oy < in_y && ox >= 0 && ox < in_x ) {
+                    for( unsigned long fx = 0; fx < side; fx++ ) {
+                        for( unsigned long  fy = 0; fy < side; fy++ ) {
+                            for( unsigned long fd = 0;fd < in_z; fd++ ) {
+                                long oy = y + (long) fy; // coordinates in the original input array coordinates
+                                long ox = x + (long) fx;
+                                if( oy >= 0 && oy < (long) in_y && ox >= 0 && ox < (long) in_x ) {
                                     // forward prop calculated: a += f.get(fx, fy, fd) * V.get(ox, oy, fd);
                                     // f.add_grad(fx, fy, fd, V.get(ox, oy, fd) * chain_grad);
                                     // V.add_grad(ox, oy, fd, f.get(fx, fy, fd) * chain_grad);
-                                    const DNN_NUMERIC in_delta = m_in_a->get( d, ox, oy, fd ) * chain_grad;
+                                    const DNN_NUMERIC in_delta = m_in_a->get( d, (unsigned long)ox, (unsigned long)oy, fd ) * chain_grad;
                                     m_dw->plusEquals( d, fx, fy, fd, in_delta );
                                     if( m_in_dw != 0 ) {
                                         const DNN_NUMERIC dw_delta = m_w->get( d, fx, fy, fd ) * chain_grad;
-                                        m_in_dw->plusEquals( d, ox, oy, fd, dw_delta );
+                                        m_in_dw->plusEquals( d, (unsigned long)ox, (unsigned long)oy, fd, dw_delta );
                                     }
                                 }
                             }
